fix(ios): don't crash the share extension when bootstrap fails #4132
LoadView returned without assigning View whenever Bootstrap returned null, and UIKit
throws on that - "NSInvalidArgumentException: attempt to insert nil object from
objects[0]". So every bootstrap failure killed the extension outright rather than
showing anything, and the crash also cut short the failure report Bootstrap had just
started sending. Reproduced on device with an induced failure: uncaught exception and
a corpse 1.3s after the failure, mid-TLS-handshake.

View is now always set. It can't be ErrorView: that is a ComputedStateView built from
IosHub, and this path runs precisely when the DI container failed to build, so
constructing it would throw inside the handler for the original failure. Its Close
would also go through UIKitExt.CloseApp -> Platform.GetCurrentUIViewController(), which
isn't guaranteed here either.

What the two can share is the screen itself, so the cat, message and Close button move
into ErrorContentView, which owns its own background and centering and takes the message
plus a close handler. ErrorView composes it pinned to its edges - the same way ShareView
hosts its children - and the controller uses it directly on the failure path, completing
the request through its own ExtensionContext with neither DI nor Platform.Init. ErrorView
loses 30 lines instead of having them duplicated, and the shared background colour stops
being copy-pasted. Being a concrete UIView subclass it joins IosShareExtAotSource, or
trimming could drop the metadata ObjC dispatch needs - on the one screen that only ever
appears once something already broke.

Verified on device: no uncaught exception, no corpse, and the sheet shows the error
screen with a working Close.
